Sync OpenBSD patchset 226:
Change the way the grid is stored, previously it was:

- a two-dimensional array of cells;
- a two-dimensional array of utf8 data;
- an array of line lengths.

Now it is a single array of a new struct grid_line each of which represents a
line and contains the length and an array of cells and an array of utf8 data.

This will make it easier to add additional per-line members, such as flags.


git-svn-id: https://tmux.svn.sourceforge.net/svnroot/tmux/trunk@1478 303bd1e2-03bb-47f1-b221-2a0928954661