Sync OpenBSD patchset 246:
Have the client pass its stdin fd to the server when identifying itself and
have the server use that rather than reopening the tty. If the fd isn't given,
use the old behaviour (so no need for a version change).

This allows tmux to be used as the shell, so also change so that when working
out the command to execute if default-command is empty (the default), tmux will
try not execute itself.


git-svn-id: https://tmux.svn.sourceforge.net/svnroot/tmux/trunk@1506 303bd1e2-03bb-47f1-b221-2a0928954661