Sync OpenBSD patchset 345:
Don't attempt to open() the tty path, rely on the client sending its stdin fd
with imsg and fatal if it doesn't, then set the FD_CLOEXEC flag in tty_init
instead of tty_open to prevent them leaking into child processes if any are
created between the two calls.

This bumps the protocol version, so the tmux server should be killed before
upgrading.


git-svn-id: https://tmux.svn.sourceforge.net/svnroot/tmux/trunk@1657 303bd1e2-03bb-47f1-b221-2a0928954661