Sync OpenBSD patchset 485:
Add an activity time for clients, like for sessions, and change session and
client lookup to pick the most recently used rather than the most recently
created - this is much more useful when used interactively and (because the
activity time is set at creation) should have no effect on source-file.

Based on a problem reported by Jan Johansson.


git-svn-id: https://tmux.svn.sourceforge.net/svnroot/tmux/trunk@1848 303bd1e2-03bb-47f1-b221-2a0928954661