Sync OpenBSD patchset 526:
There is no real standard for modifier plus function keys. Previously, tmux
output some from rxvt but in other ways did the same as xterm or other
terminals, but this is a bit inconsistent.

xterm's method is fairly sensible and we already support it (xterm-keys), so
enable it by default instead.


git-svn-id: https://tmux.svn.sourceforge.net/svnroot/tmux/trunk@1900 303bd1e2-03bb-47f1-b221-2a0928954661