Sync OpenBSD patchset 546:
Add a per-client log of status line messages displayed while that client
exists. A new message-limit session option sets the maximum number of entries
and a command, show-messages, shows the log (bound to ~ by default).

This (and prompt history) might be better as a single global log but until
there are global options it is easier for them to be per client.


git-svn-id: https://tmux.svn.sourceforge.net/svnroot/tmux/trunk@1922 303bd1e2-03bb-47f1-b221-2a0928954661