Sync OpenBSD patchset 731:
Send all three of stdin, stdout, stderr from the client to the server, so that
commands can directly make use of them. This means that load-buffer and
save-buffer can have "-" as the file to read from stdin or write to stdout.

This is a protocol version bump so the tmux server will need to be restarted
after upgrade (or an older client used).


git-svn-id: https://tmux.svn.sourceforge.net/svnroot/tmux/trunk@2204 303bd1e2-03bb-47f1-b221-2a0928954661