Minimal XIM support for utf8 input
This will only work correctly when utf8 is enabled for the window (and
the running application); ideally we would like to use it only when this
is the case.  One option would be to add another client option to enable
utf8 input.  It would be cleaner to handle this deeper in tmux,
unencoding the utf8 when the current window does not have utf8 enabled.