Fix SF bug 3354380
With epoll, if the fd is invalidated somehow, event_del will produce a message
like this:

  Epoll MOD(4) on fd 15 failed.  Old events were 6; read change was 2 (del); write change was 0 (none): Bad file descriptor

As a result, the event will not be fully deleted.  When this is done from a
bufferevent_free, the bufferevent is still removed, resulting in a crash.  This
may be considered a libevent bug, but the tmux fix is simply not to close the
fd before deleting the event.