xtmux: fix cursor misdrawing harder
The last commit only half-fixed the problem, because sometimes the corruption is due to a draw that was supposed to erase the cursor.
So, now (as a kind of hack), we clear the last known cursor position on redraw regardless of whether it should actually be there or not.
This still may not work in some cases of partial exposure, but I can't think of them right now.

Maybe this whole XOR cursor thing was a bad idea.